+++ core/examples/example_soundskrit_sst.c
@@ -21,7 +21,7 @@
 
 int main(int argc, char * argv[]) {
 
-    //                                                                                
+    //
     //                             Ms (all 1's)                                                                      dsf (default values)
     //                                  |                                                                                     |
     //                                  *                                                                                     *
@@ -72,12 +72,12 @@ int main(int argc, char * argv[]) {
     doas_t * doas_tracked = doas_construct("tracked", num_tracks);
 
     mixer_t * mixer = mixer_construct(channels);
-    stft_t * stft = stft_construct(num_channels, num_samples, num_shifts, num_bins, "hann");    
+    stft_t * stft = stft_construct(num_channels, num_samples, num_shifts, num_bins, "hann");
     scm_t * scm = scm_construct(num_channels, num_bins, alpha);
     phat_t * phat = phat_construct(num_channels, num_bins);
     fcc_t * fcc = fcc_construct(num_sources, num_channels, num_bins);
     gcc_t * gcc = gcc_construct(num_sources, num_channels, num_bins);
-    ssl_t * ssl = ssl_construct(mics, points, sample_rate, sound_speed, num_sources, num_directions);
+    ssl_t * ssl = ssl_construct(mics, points, (float)sample_rate, sound_speed, num_sources, num_directions);
     sst_t * sst = sst_construct(num_tracks, num_directions, num_pasts);
 
     msgout_t * msgout = msgout_construct("/dev/stdout");
@@ -86,10 +86,10 @@ int main(int argc, char * argv[]) {
     // Process
     //
 
-    masks_ones(masks); 
+    masks_ones(masks);
 
     while (wavin_read(wavin, hops_raw) == 0) {
-        
+
         mixer_process(mixer, hops_raw, hops_chs);
         stft_process(stft, hops_chs, freqs);
         scm_process(scm, freqs, masks, covs);
@@ -101,7 +101,7 @@ int main(int argc, char * argv[]) {
         if (strcmp(method, "fcc") == 0) {
             fcc_process(fcc, covs_phat, tdoas);
         }
-        
+
         ssl_process(ssl, tdoas, doas_potential);
         sst_process(sst, dsf, doas_potential, doas_tracked);
 
@@ -112,7 +112,7 @@ int main(int argc, char * argv[]) {
 
     }
 
-    // 
+    //
     // Free memory
     //
 
@@ -120,7 +120,7 @@ int main(int argc, char * argv[]) {
     points_destroy(points);
 
     wavin_destroy(wavin);
-    
+
     hops_destroy(hops_raw);
     hops_destroy(hops_chs);
     freqs_destroy(freqs);
@@ -131,7 +131,7 @@ int main(int argc, char * argv[]) {
     doas_destroy(doas_potential);
     dsf_destroy(dsf);
     doas_destroy(doas_tracked);
-    
+
     mixer_destroy(mixer);
     stft_destroy(stft);
     scm_destroy(scm);
